refactor trigger for flyout and integrate it with bulk actions
