+++ x-pack/plugins/spaces/public/management/view_space/view_space_roles.tsx
@@ -5,44 +5,24 @@
  * 2.0.
  */
 
-import {
-  EuiButton,
-  EuiButtonEmpty,
-  EuiButtonGroup,
-  EuiCallOut,
-  EuiComboBox,
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiFlyout,
-  EuiFlyoutBody,
-  EuiFlyoutFooter,
-  EuiFlyoutHeader,
-  EuiForm,
-  EuiFormRow,
-  EuiLink,
-  EuiSpacer,
-  EuiText,
-  EuiTitle,
-} from '@elastic/eui';
-import type { EuiComboBoxOptionOption } from '@elastic/eui';
+import { EuiFlexGroup, EuiFlexItem, EuiLink, EuiText } from '@elastic/eui';
 import type { FC } from 'react';
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import React, { useCallback, useEffect, useRef, useState } from 'react';
 
-import type { KibanaFeature, KibanaFeatureConfig } from '@kbn/features-plugin/common';
+import type { KibanaFeature } from '@kbn/features-plugin/common';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
+import { toMountPoint } from '@kbn/react-kibana-mount';
 import type { Role } from '@kbn/security-plugin-types-common';
 
+import {
+  type PrivilegesAPIClient,
+  PrivilegesRolesForm,
+  type RolesAPIClient,
+} from './component/space_assign_role_privilege_form';
 import { SpaceAssignedRolesTable } from './component/space_assigned_roles_table';
-import { useViewSpaceServices, type ViewSpaceServices } from './hooks/view_space_context_provider';
+import { useViewSpaceServices } from './hooks/view_space_context_provider';
 import type { Space } from '../../../common';
-import { FeatureTable } from '../edit_space/enabled_features/feature_table';
-
-type RolesAPIClient = ReturnType<ViewSpaceServices['getRolesAPIClient']> extends Promise<infer R>
-  ? R
-  : never;
-
-type KibanaPrivilegeBase = keyof NonNullable<KibanaFeatureConfig['privileges']>;
 
 interface Props {
   space: Space;
@@ -56,28 +36,39 @@ interface Props {
 
 // FIXME: rename to EditSpaceAssignedRoles
 export const ViewSpaceAssignedRoles: FC<Props> = ({ space, roles, features, isReadOnly }) => {
-  const [showRolesPrivilegeEditor, setShowRolesPrivilegeEditor] = useState(false);
+  // const [showRolesPrivilegeEditor, setShowRolesPrivilegeEditor] = useState(false);
   const [roleAPIClientInitialized, setRoleAPIClientInitialized] = useState(false);
   const [spaceUnallocatedRole, setSpaceUnallocatedRole] = useState<Role[]>([]);
 
   const rolesAPIClient = useRef<RolesAPIClient>();
+  const privilegesAPIClient = useRef<PrivilegesAPIClient>();
 
-  const { getRolesAPIClient, getUrlForApp } = useViewSpaceServices();
+  const {
+    getRolesAPIClient,
+    getUrlForApp,
+    getPrivilegesAPIClient,
+    overlays,
+    theme,
+    i18n: i18nStart,
+  } = useViewSpaceServices();
 
-  const resolveRolesAPIClient = useCallback(async () => {
+  const resolveAPIClients = useCallback(async () => {
     try {
-      rolesAPIClient.current = await getRolesAPIClient();
+      [rolesAPIClient.current, privilegesAPIClient.current] = await Promise.all([
+        getRolesAPIClient(),
+        getPrivilegesAPIClient(),
+      ]);
       setRoleAPIClientInitialized(true);
     } catch {
       //
     }
-  }, [getRolesAPIClient]);
+  }, [getPrivilegesAPIClient, getRolesAPIClient]);
 
   useEffect(() => {
     if (!isReadOnly) {
-      resolveRolesAPIClient();
+      resolveAPIClients();
     }
-  }, [isReadOnly, resolveRolesAPIClient]);
+  }, [isReadOnly, resolveAPIClients]);
 
   useEffect(() => {
     async function fetchAllSystemRoles() {
@@ -88,7 +79,7 @@ export const ViewSpaceAssignedRoles: FC<Props> = ({ space, roles, features, isRe
         (role) =>
           !role.metadata?._reserved &&
           role.kibana.some((privileges) => {
-            return !privileges.spaces.includes(space.id) || !privileges.spaces.includes('*');
+            return !privileges.spaces.includes(space.id) && !privileges.spaces.includes('*');
           })
       );
 
@@ -100,23 +91,35 @@ export const ViewSpaceAssignedRoles: FC<Props> = ({ space, roles, features, isRe
     }
   }, [roleAPIClientInitialized, space.id]);
 
+  const showRolesPrivilegeEditor = useCallback(
+    (defaultSelected?: Role[]) => {
+      const overlayRef = overlays.openFlyout(
+        toMountPoint(
+          <PrivilegesRolesForm
+            {...{
+              space,
+              features,
+              onSaveClick: () => overlayRef.close(),
+              closeFlyout: () => overlayRef.close(),
+              defaultSelected,
+              spaceUnallocatedRole,
+              // APIClient would have been initialized before the privilege editor is displayed
+              roleAPIClient: rolesAPIClient.current!,
+              privilegesAPIClient: privilegesAPIClient.current!,
+            }}
+          />,
+          { theme, i18n: i18nStart }
+        ),
+        {
+          size: 's',
+        }
+      );
+    },
+    [features, i18nStart, overlays, space, spaceUnallocatedRole, theme]
+  );
+
   return (
-    <>
-      {showRolesPrivilegeEditor && (
-        <PrivilegesRolesForm
-          features={features}
-          space={space}
-          closeFlyout={() => {
-            setShowRolesPrivilegeEditor(false);
-          }}
-          onSaveClick={() => {
-            setShowRolesPrivilegeEditor(false);
-          }}
-          spaceUnallocatedRole={spaceUnallocatedRole}
-          // rolesAPIClient would have been initialized before the privilege editor is displayed
-          roleAPIClient={rolesAPIClient.current!}
-        />
-      )}
+    <React.Fragment>
       <EuiFlexGroup direction="column">
         <EuiFlexItem>
           <EuiText>
@@ -138,256 +141,21 @@ export const ViewSpaceAssignedRoles: FC<Props> = ({ space, roles, features, isRe
         </EuiFlexItem>
         <EuiFlexItem>
           <SpaceAssignedRolesTable
-            isReadOnly={isReadOnly || !spaceUnallocatedRole.length}
+            isReadOnly={isReadOnly}
             assignedRoles={roles}
+            onClickBulkEdit={showRolesPrivilegeEditor}
+            onClickBulkRemove={(selectedRoles) => {
+              // TODO: add logic to remove selected roles from space
+            }}
             onAssignNewRoleClick={async () => {
               if (!roleAPIClientInitialized) {
-                await resolveRolesAPIClient();
+                await resolveAPIClients();
               }
-              setShowRolesPrivilegeEditor(true);
+              showRolesPrivilegeEditor();
             }}
           />
         </EuiFlexItem>
       </EuiFlexGroup>
-    </>
-  );
-};
-
-interface PrivilegesRolesFormProps extends Omit<Props, 'isReadOnly' | 'roles'> {
-  closeFlyout: () => void;
-  onSaveClick: () => void;
-  spaceUnallocatedRole: Role[];
-  roleAPIClient: RolesAPIClient;
-}
-
-const createRolesComboBoxOptions = (roles: Role[]): Array<EuiComboBoxOptionOption<Role>> =>
-  roles.map((role) => ({
-    label: role.name,
-    value: role,
-  }));
-
-export const PrivilegesRolesForm: FC<PrivilegesRolesFormProps> = (props) => {
-  const { onSaveClick, closeFlyout, features, roleAPIClient, spaceUnallocatedRole } = props;
-
-  const [space, setSpaceState] = useState<Partial<Space>>(props.space);
-  const [spacePrivilege, setSpacePrivilege] = useState<KibanaPrivilegeBase | 'custom'>('all');
-  const [selectedRoles, setSelectedRoles] = useState<ReturnType<typeof createRolesComboBoxOptions>>(
-    []
-  );
-  const selectedRolesHasPrivilegeConflict = useMemo(() => {
-    return selectedRoles.reduce((result, selectedRole) => {
-      // TODO: determine heuristics for role privilege conflicts
-      return result;
-    }, false);
-  }, [selectedRoles]);
-
-  const [assigningToRole, setAssigningToRole] = useState(false);
-
-  const assignRolesToSpace = useCallback(async () => {
-    try {
-      setAssigningToRole(true);
-
-      await Promise.all(
-        selectedRoles.map((selectedRole) => {
-          roleAPIClient.saveRole({ role: selectedRole.value! });
-        })
-      ).then(setAssigningToRole.bind(null, false));
-
-      onSaveClick();
-    } catch {
-      // Handle resulting error
-    }
-  }, [onSaveClick, roleAPIClient, selectedRoles]);
-
-  const getForm = () => {
-    return (
-      <EuiForm component="form" fullWidth>
-        <EuiFormRow label="Select a role(s)">
-          <EuiComboBox
-            data-test-subj="roleSelectionComboBox"
-            aria-label={i18n.translate('xpack.spaces.management.spaceDetails.roles.selectRoles', {
-              defaultMessage: 'Select role to assign to the {spaceName} space',
-              values: { spaceName: space.name },
-            })}
-            placeholder="Select roles"
-            options={createRolesComboBoxOptions(spaceUnallocatedRole)}
-            selectedOptions={selectedRoles}
-            onChange={(value) => {
-              setSelectedRoles((prevRoles) => {
-                if (prevRoles.length < value.length) {
-                  const newlyAdded = value[value.length - 1];
-
-                  const { name: spaceName } = space;
-                  if (!spaceName) {
-                    throw new Error('space state requires name!');
-                  }
-
-                  // Add kibana space privilege definition to role
-                  newlyAdded.value!.kibana.push({
-                    spaces: [spaceName],
-                    base: spacePrivilege === 'custom' ? [] : [spacePrivilege],
-                    feature: {},
-                  });
-
-                  return prevRoles.concat(newlyAdded);
-                } else {
-                  return value;
-                }
-              });
-            }}
-            fullWidth
-          />
-        </EuiFormRow>
-        <>
-          {!selectedRolesHasPrivilegeConflict && (
-            <EuiFormRow>
-              <EuiCallOut
-                color="warning"
-                iconType="iInCircle"
-                title={i18n.translate(
-                  'xpack.spaces.management.spaceDetails.roles.assign.privilegeConflictMsg.title',
-                  {
-                    defaultMessage: 'Selected roles have different privileges granted',
-                  }
-                )}
-              >
-                {i18n.translate(
-                  'xpack.spaces.management.spaceDetails.roles.assign.privilegeConflictMsg.description',
-                  {
-                    defaultMessage:
-                      'Updating the settings here in a bulk will override current individual settings.',
-                  }
-                )}
-              </EuiCallOut>
-            </EuiFormRow>
-          )}
-        </>
-        <EuiFormRow
-          helpText={i18n.translate(
-            'xpack.spaces.management.spaceDetails.roles.assign.privilegesHelpText',
-            {
-              defaultMessage:
-                'Assign the privilege you wish to grant to all present and future features across this space',
-            }
-          )}
-        >
-          <EuiButtonGroup
-            legend="select the privilege for the features enabled in this space"
-            options={[
-              {
-                id: 'all',
-                label: i18n.translate(
-                  'xpack.spaces.management.spaceDetails.roles.assign.privileges.all',
-                  {
-                    defaultMessage: 'All',
-                  }
-                ),
-              },
-              {
-                id: 'read',
-                label: i18n.translate(
-                  'xpack.spaces.management.spaceDetails.roles.assign.privileges.read',
-                  { defaultMessage: 'Read' }
-                ),
-              },
-              {
-                id: 'custom',
-                label: i18n.translate(
-                  'xpack.spaces.management.spaceDetails.roles.assign.privileges.custom',
-                  { defaultMessage: 'Customize' }
-                ),
-              },
-            ].map((privilege) => ({
-              ...privilege,
-              'data-test-subj': `${privilege.id}-privilege-button`,
-            }))}
-            color="primary"
-            idSelected={spacePrivilege}
-            onChange={(id) => setSpacePrivilege(id as KibanaPrivilegeBase | 'custom')}
-            buttonSize="compressed"
-            isFullWidth
-          />
-        </EuiFormRow>
-        {spacePrivilege === 'custom' && (
-          <EuiFormRow
-            label={i18n.translate(
-              'xpack.spaces.management.spaceDetails.roles.assign.privileges.customizeLabelText',
-              { defaultMessage: 'Customize by feature' }
-            )}
-          >
-            <>
-              <EuiText size="xs">
-                <p>
-                  <FormattedMessage
-                    id="xpack.spaces.management.spaceDetails.roles.assign.privileges.customizeDescriptionText"
-                    defaultMessage="Increase privilege levels per feature basis. Some features might be hidden by the
-                  space or affected by a global space privilege"
-                  />
-                </p>
-              </EuiText>
-              <EuiSpacer />
-              <FeatureTable space={space} features={features} onChange={setSpaceState} />
-            </>
-          </EuiFormRow>
-        )}
-      </EuiForm>
-    );
-  };
-
-  const getSaveButton = () => {
-    return (
-      <EuiButton
-        fill
-        isLoading={assigningToRole}
-        onClick={() => assignRolesToSpace()}
-        data-test-subj={'createRolesPrivilegeButton'}
-      >
-        {i18n.translate('xpack.spaces.management.spaceDetails.roles.assignRoleButton', {
-          defaultMessage: 'Assign roles',
-        })}
-      </EuiButton>
-    );
-  };
-
-  return (
-    <EuiFlyout onClose={closeFlyout} size="s">
-      <EuiFlyoutHeader hasBorder>
-        <EuiTitle size="m">
-          <h2>
-            {i18n.translate('xpack.spaces.management.spaceDetails.roles.assign.privileges.custom', {
-              defaultMessage: 'Assign role to {spaceName}',
-              values: { spaceName: space.name },
-            })}
-          </h2>
-        </EuiTitle>
-        <EuiSpacer size="s" />
-        <EuiText size="s">
-          <p>
-            <FormattedMessage
-              id="xpack.spaces.management.spaceDetails.privilegeForm.heading"
-              defaultMessage="Roles will be granted access to the current space according to their default privileges. Use the &lsquo;Customize&rsquo; option to override default privileges."
-            />
-          </p>
-        </EuiText>
-      </EuiFlyoutHeader>
-      <EuiFlyoutBody>{getForm()}</EuiFlyoutBody>
-      <EuiFlyoutFooter>
-        <EuiFlexGroup justifyContent="spaceBetween">
-          <EuiFlexItem grow={false}>
-            <EuiButtonEmpty
-              iconType="cross"
-              onClick={closeFlyout}
-              flush="left"
-              data-test-subj={'cancelRolesPrivilegeButton'}
-            >
-              {i18n.translate('xpack.spaces.management.spaceDetails.roles.cancelRoleButton', {
-                defaultMessage: 'Cancel',
-              })}
-            </EuiButtonEmpty>
-          </EuiFlexItem>
-          <EuiFlexItem grow={false}>{getSaveButton()}</EuiFlexItem>
-        </EuiFlexGroup>
-      </EuiFlyoutFooter>
-    </EuiFlyout>
+    </React.Fragment>
   );
 };
